Added Teams!
Still a basic version of teams.
From now on only buildings and characters that belong to the player can
be selected (resources are considered neutral and available for
everyone).
Added a bool on BP_RTS_Character that allows editor-placed characters to
look for a player owned team. Maybe this is also needed for buildings?