Update to 102 + HUD Idea
I've added a new idea for hud until we find something worth using.
placing a building has been removed.
This is just a proof of concept and will need more work.
By default there is one button on the hud printing Hello when clicked,
if you have selected a character (of your own team) it adds a second
button printing Yeah when clicked.